+++ lex.config.ts
@@ -3,5 +3,5 @@ import { defineLexiconConfig } from "@atcute/lex-cli";
 
 export default defineLexiconConfig({
   files: ["lexicons/**/*.json"],
-  outdir: "src/lexicons/",
+  outdir: "lexicons/",
 });